+++ .github/workflows/auto-generated.yml
@@ -1,0 +1,50 @@
+# This is a basic workflow to help you get started with Actions
+
+name: 🔨 Auto Generating Tasks
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push or pull request events but only for the main branch
+  push:
+    branches: [main]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+concurrency:
+  group: commit-work
+  cancel-in-progress: false
+
+env:
+  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+jobs:
+  linking:
+    runs-on: ubuntu-latest
+    name: Linking markdown files
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - name: 📦 Checkout Repository
+        uses: actions/checkout@v3
+
+
+
+      # Runs a single command using the runners shell
+      - name: 📓 Create Readmes
+        uses: DaanV2/Markdown-Action-Create-Indexes@v2.0.0
+        with:
+          folder: ${{github.workspace}}/Benchmark
+          filename: README.md
+
+      - name: 📖 Commit changes
+        continue-on-error: true
+        run: |
+          cd $GITHUB_WORKSPACE
+          git config --global user.email "github-action@github.com"
+          git config --global user.name "Tasha"
+          git add .
+          git commit -m "auto: Generated Readmes"
+
+      - name: 🚚 Push
+        continue-on-error: true
+        run: git push
