+++ Benchmark/Dictionary/Class.cs
@@ -1,0 +1,7 @@
+﻿namespace Benchmark.Dictionary;
+
+public class DictonaryTestClass {
+    public String Data1 { get; set; }
+    public String Data2 { get; set; }
+    public String Data3 { get; set; }
+}
