+++ Benchmark/Mechanics/VersionMask.cs
@@ -4,7 +4,7 @@ using BenchmarkDotNet.Engines;
 using DaanV2.UUID;
 using Version = DaanV2.UUID.Version;
 
-namespace Benchmark;
+namespace Benchmark.Mechanics;
 
 [MemoryDiagnoser]
 [SimpleJob(RunStrategy.Throughput, id: "Version masking, as vector or just bytes?", iterationCount: 50, invocationCount: 5000000)]
@@ -21,7 +21,7 @@ public partial class VersionMask {
     public static Vector128<Byte> versionMask = Format.VersionVariantMaskNot(Version.V4, Variant.V1);
     public static Vector128<Byte> versionOverlay = Format.VersionVariantOverlayer(Version.V4, Variant.V1);
 
-    public Byte[] Data { get; set; }
+    public Byte[] Data { get; set; } = Array.Empty<Byte>();
     public Version Ver { get; set; }
     public Variant Var { get; set; }
 
@@ -59,18 +59,18 @@ public partial class VersionMask {
         return new UUID(result);
     }
 
-    [Benchmark(Description = "Using bytes and constant values", Baseline =true)]
+    [Benchmark(Description = "Using bytes and constant values", Baseline = true)]
     public UUID Bytes() {
         Span<Byte> data = this.Data.AsSpan();
 
         UInt32 maskVer = Version.V4.GetMask();
         Byte b = data[Format.VERSION_BYTE_INDEX];
-        b = (Byte)((b & maskVer) | (UInt32)Version.V4);
+        b = (Byte)(b & maskVer | (UInt32)Version.V4);
         data[Format.VERSION_BYTE_INDEX] = b;
 
         UInt32 maskVar = Variant.V1.GetMask();
         b = data[Format.VARIANT_BYTE_INDEX];
-        b = (Byte)((b & maskVar) | (UInt32)Variant.V1);
+        b = (Byte)(b & maskVar | (UInt32)Variant.V1);
         data[Format.VARIANT_BYTE_INDEX] = b;
 
         return new UUID(data);
