--- Benchmark/Reports/Reports.md
@@ -1,70 +1,0 @@
-# Benchmark reports
-
-- [Benchmark reports](#benchmark-reports)
-  - [Array](#array)
-  - [Dictionary](#dictionary)
-  - [List](#list)
-  - [Generation](#generation)
-    - [AllVersions](#allversions)
-    - [BatchAllVersions](#batchallversions)
-    - [GenerationBatchV4](#generationbatchv4)
-    - [GenerationV4](#generationv4)
-    - [StringBatchGenerationV4](#stringbatchgenerationv4)
-    - [StringGenerationV4](#stringgenerationv4)
-  - [Serialization](#serialization)
-    - [JsonDeserializationTest](#jsondeserializationtest)
-  - [Other](#other)
-    - [VersionMask](#versionmask)
-
-
-## Array
-
-[lookup benchmark](./results/Benchmark.ArrayTest-report-github.md)
-
-## Dictionary
-
-[lookup benchmark](./results/Benchmark.DictionaryTest-report-github.md)
-
-## List
-
-[lookup benchmark](./results/Benchmark.ListTest-report-github.md)
-
-## Generation
-
-### AllVersions
-
-[report](./results/Benchmark.Generation.AllVersions-report-github.md)
-
-### BatchAllVersions
-
-[report](./results/Benchmark.Generation.BatchAllVersions-report-github.md)
-
-### GenerationBatchV4
-
-[report](./results/Benchmark.Generation.GenerationBatchV4-report-github.md)
-
-### GenerationV4
-
-[report](./results/Benchmark.Generation.GenerationV4-report-github.md)
-
-### StringBatchGenerationV4
-
-[report](./results/Benchmark.Generation.StringBatchGenerationV4-report-github.md)
-
-### StringGenerationV4
-
-[report](./results/Benchmark.Generation.StringGenerationV4-report-github.md)
-
-## Serialization
-
-### JsonDeserializationTest
-
-[report](./results/Benchmark.Serialization.JsonDeserializationTest-report-github.md)
-
-## Other
-
-### VersionMask
-
-[stamping version](./results/Benchmark.VersionMask-report-github.md)
-
-
